Moved WebSocketClient under websocket package
